requirements.txt
cffi==1.16.0
contourpy==1.2.1
cycler==0.12.1
docstring_parser==0.16
evdev==1.7.1
fonttools==4.53.1
kiwisolver==1.4.5
markdown-it-py==3.0.0
matplotlib==3.9.1
mdurl==0.1.2
numpy==1.26.4
packaging==24.1
pillow==10.4.0
PyAudio==0.2.14
pycparser==2.22
Pygments==2.18.0
pynput==1.7.7
pyparsing==3.1.2
pyrtlsdr==0.3.0
python-dateutil==2.9.0.post0
python-xlib==0.33
rich==13.7.1
ruff==0.5.5
scipy==1.14.0
shtab==1.7.1
six==1.16.0
sounddevice==0.4.7
typing_extensions==4.12.2
tyro==0.8.5
